PR feedback, clarify comments - Share more code across coreclr + mono
