+++ src/coreclr/System.Private.CoreLib/src/System/Exception.CoreCLR.cs
@@ -7,7 +7,6 @@ using System.Reflection;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 using System.Runtime.Serialization;
-using System.Text;
 
 namespace System
 {
@@ -323,40 +322,9 @@ namespace System
                 _remoteStackTraceString, _ipForWatsonBuckets, _watsonBuckets);
         }
 
-        [StackTraceHidden]
-        internal void SetCurrentStackTrace()
-        {
-            if (!CanSetRemoteStackTrace())
-            {
-                return; // early-exit
-            }
-
-            // Store the current stack trace into the "remote" stack trace, which was originally introduced to support
-            // remoting of exceptions cross app-domain boundaries, and is thus concatenated into Exception.StackTrace
-            // when it's retrieved.
-            var sb = new StringBuilder(256);
-            new StackTrace(fNeedFileInfo: true).ToString(System.Diagnostics.StackTrace.TraceFormat.TrailingNewLine, sb);
-            sb.AppendLine(SR.Exception_EndStackTraceFromPreviousThrow);
-            _remoteStackTraceString = sb.ToString();
-        }
-
-        [StackTraceHidden]
-        internal void SetRemoteStackTrace(string stackTrace)
-        {
-            if (!CanSetRemoteStackTrace())
-            {
-                return; // early-exit
-            }
-
-            // Store the provided text into the "remote" stack trace, following the same format SetCurrentStackTrace
-            // would have generated.
-            _remoteStackTraceString = stackTrace + Environment.NewLineConst + SR.Exception_EndStackTraceFromPreviousThrow + Environment.NewLineConst;
-        }
-
         // Returns true if setting the _remoteStackTraceString field is legal, false if not (immutable exception).
         // A false return value means the caller should early-exit the operation.
         // Can also throw InvalidOperationException if a stack trace is already set or if object has been thrown.
-        [StackTraceHidden]
         private bool CanSetRemoteStackTrace()
         {
             // If this is a preallocated singleton exception, silently skip the operation,
